requirements.txt
pytest==9.0.2
requests==2.33.1
python-dotenv==1.2.2